split up election-timer into additional files
