+++ src/server/election-timer/election-timeout.ts
@@ -1,0 +1,8 @@
+export type ElectionTimeout = number;
+
+export type ElectionTimeoutInterval = [number, number];
+
+// The Raft paper recommends an election timeout
+// interval of 150–300 milliseconds:
+// > *§9.3 "...We recommend using a conservative election timeout such as 150–300ms..."
+export const DEFAULT_ELECTION_TIMEOUT_INTERVAL: ElectionTimeoutInterval = [150, 300];
