+++ src/server/election-timer/factory.ts
@@ -1,11 +1,16 @@
 import {
     ElectionTimer,
-    ElectionTimeoutChooser,
     IElectionTimerOptions,
-    IElectionTimer,
+    IElectionTimer
+} from './election-timer';
+
+import { DEFAULT_ELECTION_TIMEOUT_INTERVAL } from './election-timeout';
+
+import {
+    ElectionTimeoutChooser,
     IElectionTimeoutChooserOptions,
     IElectionTimeoutChooser
-} from './election-timer';
+} from './election-timeout-chooser';
 
 export function createElectionTimer(options: IElectionTimerOptions = {}): IElectionTimer {
     const timeoutChooser = options.timeoutChooser ? options.timeoutChooser : createElectionTimeoutChooser();
@@ -13,5 +18,6 @@ export function createElectionTimer(options: IElectionTimerOptions = {}): IElect
 }
 
 export function createElectionTimeoutChooser(options: IElectionTimeoutChooserOptions = {}): IElectionTimeoutChooser {
-    return new ElectionTimeoutChooser(options);
+    const interval = options.interval ? options.interval : DEFAULT_ELECTION_TIMEOUT_INTERVAL;
+    return new ElectionTimeoutChooser(Object.assign({}, options, { interval }));
 }
