+++ src/server/election-timer/index.ts
@@ -1,2 +1,4 @@
 export * from './election-timer';
+export * from './election-timeout';
+export * from './election-timeout-chooser';
 export * from './factory';
